+++ src/api/index.js
@@ -9,4 +9,4 @@ export * from 'src/api/sync';
 export * from 'src/api/calls';
 export * from 'src/api/notificationSettings';
 export * from 'src/api/callNotes';
-export * from 'src/api/chat';
+export * from 'src/api/messages';
