--- src/containers/ChatContainer.js
@@ -1,5 +1,0 @@
-import { connect } from 'react-redux';
-import Chat from 'src/views/Chat';
-
-
-export default connect()(Chat);
