+++ src/containers/MessagesContainer.js
@@ -1,0 +1,5 @@
+import { connect } from 'react-redux';
+import Messages from 'src/views/Messages';
+
+
+export default connect()(Messages);
