+++ src/routes.js
@@ -43,7 +43,7 @@ import CallNoteSavedContainer from 'src/containers/CallNoteSavedContainer';
 import CallNoteListContainer from 'src/containers/CallNoteListContainer';
 import CallNoteDetailContainer from 'src/containers/CallNoteDetailContainer';
 
-import Chat from 'src/containers/ChatContainer';
+import Messages from 'src/containers/MessagesContainer';
 
 import ProfileSettingsContainer from 'src/containers/ProfileSettingsContainer';
 
@@ -91,7 +91,7 @@ export default fromPairs([
   [constants.ROUTE_CALL_NOTE_LIST, CallNoteListContainer],
   [constants.ROUTE_CALL_NOTE_DETAIL, CallNoteDetailContainer],
 
-  [constants.ROUTE_CHAT, Chat],
+  [constants.ROUTE_CHAT, Messages],
 
   [constants.ROUTE_PROFILE_SETTINGS, ProfileSettingsContainer],
 ]);
